.claude/skills/feishu-send/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
requests-toolbelt>=1.0.0
